--- common/host.go
@@ -1,6 +1,0 @@
-package common
-
-var (
-    // HostIPs contain Docker host ip address for monitor
-    HostIPs = []string{"localhost"}
-)
