+++ internal/conf/stats.go
@@ -1,0 +1,5 @@
+package conf
+
+const (
+	DefaultReadLength = 15
+)
