+++ internal/conf/time.go
@@ -1,0 +1,9 @@
+package conf
+
+import "time"
+
+const (
+	DefaultCollectDuration = 58 * time.Second
+	DefaultCollectTimeOut  = DefaultCollectDuration + 10*time.Second
+	DefaultMaxTimeoutTimes = 5
+)
