+++ internal/util/events.go
@@ -1,4 +1,4 @@
-package container
+package util
 
 import (
 	"sync"
