+++ script/build.sh
@@ -1,4 +1,3 @@
 #!/bin/bash
 echo "IP value is not given, default localhsot, please consider your local test env, and gin is in debug mode"
-# nothing 
-go build -tags=jsoniter
+go build -tags=jsoniter  -o monitor ../cmd/monitor/monitor.go
